move events to their own directory
